Codebase converted to Typescript
