+++ jest.config.js
@@ -1,0 +1,8 @@
+/*
+## We are currently using JEST defaults
+*/
+const { defaults: tsjPreset } = require("ts-jest/presets");
+
+module.exports = {
+  transform: tsjPreset.transform,
+};
