+++ package.json
@@ -7,7 +7,7 @@
     "start": "npm run dev",
     "dev": "node src/app.js | pino-pretty",
     "format": "prettier --write .",
-    "test:lint": "eslint src",
+    "test:lint": "eslint src/**/*.ts",
     "test:format": "prettier --check .",
     "test:unit": "jest",
     "test": "npm run test:unit && npm run test:lint && npm run test:format",
@@ -18,14 +18,23 @@
   "devDependencies": {
     "@jest/globals": "^26.6.2",
     "@kth/eslint-config-kth": "^1.0.10",
+    "@swc-node/jest": "^1.3.2",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^15.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
     "eslint": "^7.29.0",
-    "jest": "^26.6.3",
+    "jest": "^27.0.0",
     "pino-pretty": "^4.5.0",
-    "prettier": "^2.3.2"
+    "prettier": "^2.3.2",
+    "ts-jest": "^27.0.7",
+    "ts-node-dev": "^1.1.6",
+    "typescript": "^4.2.4"
   },
   "dependencies": {
     "@kth/canvas-api": "^3.0.2",
     "@kth/reqvars": "^2.0.1",
+    "@swc/core": ">=1.2.0",
     "dotenv": "^8.2.0",
     "fast-csv": "^4.3.6",
     "got": "^11.8.1",
@@ -33,6 +42,7 @@
     "ldapts": "^2.11.0",
     "memoizee": "^0.4.15",
     "pino": "^6.11.1",
-    "skog": "^2.1.0"
+    "skog": "^2.1.0",
+    "ts-node": "^10.4.0"
   }
 }
